Redis/requirements.txt
boto3==1.26.50
botocore==1.29.50
jmespath==1.0.1
python-dateutil==2.8.2
PyYAML==6.0
s3transfer==0.6.0
six==1.16.0
types-ujson==4.2.1
ujson==5.7.0
urllib3==1.26.14
